+++ previous-versions/common/pkg/models/operations/configuresupplementaldata.go
@@ -124,11 +124,14 @@ func (o *ConfigureSupplementalDataRequest) GetPlatformKey() string {
 }
 
 type ConfigureSupplementalDataResponse struct {
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *ConfigureSupplementalDataResponse) GetContentType() string {