+++ previous-versions/common/pkg/models/operations/getconnection.go
@@ -28,12 +28,15 @@ func (o *GetConnectionRequest) GetConnectionID() string {
 
 type GetConnectionResponse struct {
 	// OK
-	Connection  *shared.Connection
+	Connection *shared.Connection
+	// HTTP response content type for this operation
 	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *GetConnectionResponse) GetConnection() *shared.Connection {