+++ previous-versions/common/RELEASES.md
@@ -353,3 +353,13 @@ Based on:
 - [go v0.26.0] previous-versions/common
 ### Releases
 - [Go v0.26.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/common/v0.26.0 - previous-versions/common
+
+## 2023-10-04 00:05:03
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.93.2 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.27.0] previous-versions/common
+### Releases
+- [Go v0.27.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/common/v0.27.0 - previous-versions/common